+++ test/integration/cli-nics.test.js
@@ -41,7 +41,7 @@ test('triton instance nics', OPTS, function (tt) {
     h.printConfig(tt);
 
     tt.test('  cleanup existing inst with alias ' + INST_ALIAS, function (t) {
-        h.deleteTestInst(t, INST_ALIAS, function (err) {
+        h.deleteTestInst(t, INST_ALIAS, function onDelete(err) {
             t.ifErr(err);
             t.end();
         });
@@ -49,8 +49,10 @@ test('triton instance nics', OPTS, function (tt) {
 
     tt.test('  setup: triton instance create', function (t) {
         h.createTestInst(t, INST_ALIAS, {}, function onInst(err, instId) {
-            if (h.ifErr(t, err, 'triton instance create'))
-                return t.end();
+            if (h.ifErr(t, err, 'triton instance create')) {
+                t.end();
+                return;
+            }
 
             t.ok(instId, 'created instance ' + instId);
             INST = instId;
@@ -61,8 +63,10 @@ test('triton instance nics', OPTS, function (tt) {
 
     tt.test('  setup: find network for tests', function (t) {
         h.triton('network list -j', function onNetworks(err, stdout) {
-            if (h.ifErr(t, err, 'triton network list'))
-                return t.end();
+            if (h.ifErr(t, err, 'triton network list')) {
+                t.end();
+                return;
+            }
 
             NETWORK = JSON.parse(stdout.trim().split('\n')[0]);
             t.ok(NETWORK, 'NETWORK');
@@ -74,9 +78,11 @@ test('triton instance nics', OPTS, function (tt) {
     tt.test('  triton instance nic create', function (t) {
         var cmd = 'instance nic create -j -w ' + INST + ' ' + NETWORK.id;
 
-        h.triton(cmd, function (err, stdout, stderr) {
-            if (h.ifErr(t, err, 'triton instance nic create'))
-                return t.end();
+        h.triton(cmd, function onTriton(err, stdout, stderr) {
+            if (h.ifErr(t, err, 'triton instance nic create')) {
+                t.end();
+                return;
+            }
 
             NIC = JSON.parse(stdout);
             t.ok(NIC, 'created NIC: ' + stdout.trim());
@@ -88,9 +94,11 @@ test('triton instance nics', OPTS, function (tt) {
     tt.test('  triton instance nic get', function (t) {
         var cmd = 'instance nic get ' + INST + ' ' + NIC.mac;
 
-        h.triton(cmd, function (err, stdout, stderr) {
-            if (h.ifErr(t, err, 'triton instance nic get'))
-                return t.end();
+        h.triton(cmd, function onTriton(err, stdout, stderr) {
+            if (h.ifErr(t, err, 'triton instance nic get')) {
+                t.end();
+                return;
+            }
 
             var obj = JSON.parse(stdout);
             t.equal(obj.mac, NIC.mac, 'nic MAC is correct');
@@ -104,9 +112,11 @@ test('triton instance nics', OPTS, function (tt) {
     tt.test('  triton instance nic list', function (t) {
         var cmd = 'instance nic list ' + INST;
 
-        h.triton(cmd, function (err, stdout, stderr) {
-            if (h.ifErr(t, err, 'triton instance nic list'))
-                return t.end();
+        h.triton(cmd, function onTriton(err, stdout, stderr) {
+            if (h.ifErr(t, err, 'triton instance nic list')) {
+                t.end();
+                return;
+            }
 
             var nics = stdout.trim().split('\n');
             t.ok(nics[0].match(/IP\s+MAC\s+STATE\s+NETWORK/), 'nic list' +
@@ -115,7 +125,7 @@ test('triton instance nics', OPTS, function (tt) {
 
             t.ok(nics.length >= 1, 'triton nic list expected nic num');
 
-            var testNics = nics.filter(function (nic) {
+            var testNics = nics.filter(function doFilter(nic) {
                 return nic.match(NIC.mac);
             });
 
@@ -128,17 +138,19 @@ test('triton instance nics', OPTS, function (tt) {
     tt.test('  triton instance nic list -j', function (t) {
         var cmd = 'instance nic list -j ' + INST;
 
-        h.triton(cmd, function (err, stdout, stderr) {
-            if (h.ifErr(t, err, 'triton instance nic list'))
-                return t.end();
+        h.triton(cmd, function onTriton(err, stdout, stderr) {
+            if (h.ifErr(t, err, 'triton instance nic list')) {
+                t.end();
+                return;
+            }
 
-            var nics = stdout.trim().split('\n').map(function (line) {
+            var nics = stdout.trim().split('\n').map(function doParse(line) {
                 return JSON.parse(line);
             });
 
             t.ok(nics.length >= 1, 'triton nic list expected nic num');
 
-            var testNics = nics.filter(function (nic) {
+            var testNics = nics.filter(function doFilter(nic) {
                 return nic.mac === NIC.mac;
             });
 
@@ -150,11 +162,13 @@ test('triton instance nics', OPTS, function (tt) {
 
     tt.test(' triton instance nic list mac=<...>', function (t) {
         var cmd = 'instance nic list -j ' + INST + ' mac=' + NIC.mac;
-        h.triton(cmd, function (err, stdout, stderr) {
-            if (h.ifErr(t, err))
-                return t.end();
+        h.triton(cmd, function onTriton(err, stdout, stderr) {
+            if (h.ifErr(t, err)) {
+                t.end();
+                return;
+            }
 
-            var nics = stdout.trim().split('\n').map(function (str) {
+            var nics = stdout.trim().split('\n').map(function doParse(str) {
                 return JSON.parse(str);
             });
 
@@ -169,11 +183,13 @@ test('triton instance nics', OPTS, function (tt) {
     tt.test(' triton nic list mac=<...>', function (t) {
         var cmd = 'instance nic list -j ' + INST + ' mac=' + NIC.mac;
 
-        h.triton(cmd, function (err, stdout, stderr) {
-            if (h.ifErr(t, err))
-                return t.end();
+        h.triton(cmd, function doTriton(err, stdout, stderr) {
+            if (h.ifErr(t, err)) {
+                t.end();
+                return;
+            }
 
-            var nics = stdout.trim().split('\n').map(function (str) {
+            var nics = stdout.trim().split('\n').map(function doParse(str) {
                 return JSON.parse(str);
             });
 
@@ -188,9 +204,11 @@ test('triton instance nics', OPTS, function (tt) {
     tt.test('  triton instance nic delete', function (t) {
         var cmd = 'instance nic delete --force ' + INST + ' ' + NIC.mac;
 
-        h.triton(cmd, function (err, stdout, stderr) {
-            if (h.ifErr(t, err, 'triton instance nic delete'))
-                return t.end();
+        h.triton(cmd, function doTriton(err, stdout, stderr) {
+            if (h.ifErr(t, err, 'triton instance nic delete')) {
+                t.end();
+                return;
+            }
 
             t.ok(stdout.match('Deleted NIC ' + NIC.mac, 'deleted nic'));
 
@@ -202,9 +220,11 @@ test('triton instance nics', OPTS, function (tt) {
         var cmd = 'instance nic create -j -w ' + INST + ' ipv4_uuid=' +
             NETWORK.id;
 
-        h.triton(cmd, function (err, stdout, stderr) {
-            if (h.ifErr(t, err, 'triton instance nic create'))
-                return t.end();
+        h.triton(cmd, function doTriton(err, stdout, stderr) {
+            if (h.ifErr(t, err, 'triton instance nic create')) {
+                t.end();
+                return;
+            }
 
             NIC2 = JSON.parse(stdout);
 
@@ -215,9 +235,11 @@ test('triton instance nics', OPTS, function (tt) {
     tt.test('  triton instance nic with ip get', function (t) {
         var cmd = 'instance nic get ' + INST + ' ' + NIC2.mac;
 
-        h.triton(cmd, function (err, stdout, stderr) {
-            if (h.ifErr(t, err, 'triton instance nic get'))
-                return t.end();
+        h.triton(cmd, function onTriton(err, stdout, stderr) {
+            if (h.ifErr(t, err, 'triton instance nic get')) {
+                t.end();
+                return;
+            }
 
             var obj = JSON.parse(stdout);
             t.equal(obj.mac, NIC2.mac, 'nic MAC is correct');
@@ -231,9 +253,11 @@ test('triton instance nics', OPTS, function (tt) {
     tt.test('  triton instance nic with ip delete', function (t) {
         var cmd = 'instance nic delete --force ' + INST + ' ' + NIC2.mac;
 
-        h.triton(cmd, function (err, stdout, stderr) {
-            if (h.ifErr(t, err, 'triton instance nic with ip delete'))
-                return t.end();
+        h.triton(cmd, function onTriton(err, stdout, stderr) {
+            if (h.ifErr(t, err, 'triton instance nic with ip delete')) {
+                t.end();
+                return;
+            }
 
             t.ok(stdout.match('Deleted NIC ' + NIC2.mac, 'deleted nic'));